Consolidate player control logic
Moves the control logic into a class for the skins to use rather than duplicating that functionality
