+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/skins/AudioPlayerSkin.kt
@@ -1,9 +1,5 @@
 package org.wycliffeassociates.otter.jvm.controls.skins
 
-import com.github.thomasnield.rxkotlinfx.observeOnFx
-import io.reactivex.Observable
-import io.reactivex.disposables.Disposable
-import javafx.application.Platform
 import javafx.fxml.FXML
 import javafx.fxml.FXMLLoader
 import javafx.scene.Node
@@ -12,12 +8,12 @@ import javafx.scene.control.SkinBase
 import javafx.scene.control.Slider
 import javafx.scene.layout.HBox
 import org.kordamp.ikonli.javafx.FontIcon
-import org.wycliffeassociates.otter.common.device.AudioPlayerEvent
 import org.wycliffeassociates.otter.jvm.controls.AudioPlayerNode
+import org.wycliffeassociates.otter.jvm.controls.controllers.AudioPlayerController
 import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
-import java.util.concurrent.TimeUnit
+import tornadofx.*
 
-class AudioPlayerSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlayerNode>(player) {
+class AudioPlayerSkin(private val playerNode: AudioPlayerNode) : SkinBase<AudioPlayerNode>(playerNode) {
 
     private val PLAY_ICON = FontIcon("fa-play")
     private val PAUSE_ICON = FontIcon("fa-pause")
@@ -31,9 +27,7 @@ class AudioPlayerSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlaye
     @FXML
     lateinit var audioSlider: Slider
 
-    private var disposable: Disposable? = null
-    private var dragging = false
-    private var resumeAfterDrag = false
+    lateinit var audioController: AudioPlayerController
 
     init {
         loadFXML()
@@ -41,77 +35,28 @@ class AudioPlayerSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlaye
     }
 
     private fun initializeControl() {
+        audioController = AudioPlayerController(playerNode.audioPlayerProperty.value, audioSlider)
         playBtn.setOnMouseClicked {
-            toggle()
+            audioController.toggle()
         }
-        if (player.sourceAvailable) {
-            sourceMissing.visibleProperty().set(false)
-            audioPlayer.visibleProperty().set(true)
-        } else {
-            sourceMissing.visibleProperty().set(true)
-            audioPlayer.visibleProperty().set(false)
-        }
-        audioSlider.value = 0.0
-        audioSlider.setOnDragDetected {
-            if (player.isPlaying) {
-                resumeAfterDrag = true
-                toggle()
-            }
-            dragging = true
-        }
-        audioSlider.setOnMouseReleased {
-            player.seek(audioSlider.value.toFloat() / 100F)
-            if (dragging) {
-                dragging = false
-                if (resumeAfterDrag) {
-                    toggle()
-                    resumeAfterDrag = false
-                }
-            }
-        }
-        player.isPlayingProperty.onChangeAndDoNow {
+        audioController.isPlayingProperty.onChangeAndDoNow {
             if (it == true) {
                 playBtn.graphicProperty().set(PAUSE_ICON)
             } else {
                 playBtn.graphicProperty().set(PLAY_ICON)
             }
         }
-    }
-
-    private fun startProgressUpdate(): Disposable {
-        return Observable
-            .interval(20, TimeUnit.MILLISECONDS)
-            .observeOnFx()
-            .subscribe {
-                if (player.isPlaying && !audioSlider.isValueChanging && !dragging) {
-                    audioSlider.value = player.playbackPosition()
-                }
+        playerNode.audioPlayerProperty.onChange { player ->
+            player?.let {
+                audioController.load(it)
             }
-    }
-
-    private fun toggle() {
-        disposable?.dispose()
-        if (player.isPlaying) {
-            player.pause()
+        }
+        if (playerNode.sourceAvailable) {
+            sourceMissing.visibleProperty().set(false)
+            audioPlayer.visibleProperty().set(true)
         } else {
-            disposable = startProgressUpdate()
-            player.play()
-            player.player?.addEventListener {
-                if (
-                    it == AudioPlayerEvent.PAUSE ||
-                    it == AudioPlayerEvent.STOP ||
-                    it == AudioPlayerEvent.COMPLETE
-                ) {
-                    disposable?.dispose()
-                    Platform.runLater {
-                        player.isPlayingProperty.set(false)
-                        if (it == AudioPlayerEvent.COMPLETE) {
-                            audioSlider.value = 0.0
-                        }
-                    }
-                }
-            }
-            player.isPlayingProperty.set(true)
+            sourceMissing.visibleProperty().set(true)
+            audioPlayer.visibleProperty().set(false)
         }
     }
