+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/skins/SourceAudioSkin.kt
@@ -1,9 +1,5 @@
 package org.wycliffeassociates.otter.jvm.controls.skins
 
-import com.github.thomasnield.rxkotlinfx.observeOnFx
-import io.reactivex.Observable
-import io.reactivex.disposables.Disposable
-import javafx.application.Platform
 import javafx.fxml.FXML
 import javafx.fxml.FXMLLoader
 import javafx.scene.Node
@@ -11,12 +7,12 @@ import javafx.scene.control.Button
 import javafx.scene.control.SkinBase
 import javafx.scene.control.Slider
 import org.kordamp.ikonli.javafx.FontIcon
-import org.wycliffeassociates.otter.common.device.AudioPlayerEvent
 import org.wycliffeassociates.otter.jvm.controls.AudioPlayerNode
+import org.wycliffeassociates.otter.jvm.controls.controllers.AudioPlayerController
 import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
-import java.util.concurrent.TimeUnit
+import tornadofx.*
 
-class SourceAudioSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlayerNode>(player) {
+class SourceAudioSkin(private val playerNode: AudioPlayerNode) : SkinBase<AudioPlayerNode>(playerNode) {
 
     private val PLAY_ICON = FontIcon("fa-play")
     private val PAUSE_ICON = FontIcon("fa-pause")
@@ -26,9 +22,7 @@ class SourceAudioSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlaye
     @FXML
     lateinit var audioSlider: Slider
 
-    private var disposable: Disposable? = null
-    private var dragging = false
-    private var resumeAfterDrag = false
+    lateinit var audioController: AudioPlayerController
 
     init {
         loadFXML()
@@ -36,70 +30,21 @@ class SourceAudioSkin(private val player: AudioPlayerNode) : SkinBase<AudioPlaye
     }
 
     private fun initializeControl() {
+        audioController = AudioPlayerController(playerNode.audioPlayerProperty.value, audioSlider)
         playBtn.setOnMouseClicked {
-            toggle()
+            audioController.toggle()
         }
-        audioSlider.value = 0.0
-        audioSlider.setOnDragDetected {
-            if (player.isPlaying) {
-                resumeAfterDrag = true
-                toggle()
-            }
-            dragging = true
-        }
-        audioSlider.setOnMouseReleased {
-            player.seek(audioSlider.value.toFloat() / 100F)
-            if (dragging) {
-                dragging = false
-                if (resumeAfterDrag) {
-                    toggle()
-                    resumeAfterDrag = false
-                }
-            }
-        }
-        player.isPlayingProperty.onChangeAndDoNow {
+        audioController.isPlayingProperty.onChangeAndDoNow {
             if (it == true) {
                 playBtn.graphicProperty().set(PAUSE_ICON)
             } else {
                 playBtn.graphicProperty().set(PLAY_ICON)
             }
         }
-    }
-
-    private fun startProgressUpdate(): Disposable {
-        return Observable
-            .interval(20, TimeUnit.MILLISECONDS)
-            .observeOnFx()
-            .subscribe {
-                if (player.isPlaying && !audioSlider.isValueChanging && !dragging) {
-                    audioSlider.value = player.playbackPosition()
-                }
+        playerNode.audioPlayerProperty.onChange { player ->
+            player?.let {
+                audioController.load(it)
             }
-    }
-
-    private fun toggle() {
-        disposable?.dispose()
-        if (player.isPlaying) {
-            player.pause()
-        } else {
-            disposable = startProgressUpdate()
-            player.play()
-            player.player?.addEventListener {
-                if (
-                    it == AudioPlayerEvent.PAUSE ||
-                    it == AudioPlayerEvent.STOP ||
-                    it == AudioPlayerEvent.COMPLETE
-                ) {
-                    disposable?.dispose()
-                    Platform.runLater {
-                        player.isPlayingProperty.set(false)
-                        if (it == AudioPlayerEvent.COMPLETE) {
-                            audioSlider.value = 0.0
-                        }
-                    }
-                }
-            }
-            player.isPlayingProperty.set(true)
         }
     }
